impr: re-organize ryot_sprite systems
