+++ crates/ryot_sprites/src/lib.rs
@@ -5,15 +5,11 @@
 use bevy_ecs::prelude::SystemSet;
 
 pub mod animations;
+pub mod loading;
 pub mod material;
-pub mod meshes;
-pub mod params;
-
 // TODO: make this better module, break it down
 pub mod sprite_animations;
-
-// TODO: make this better module, break it down
-pub mod sprites;
+pub mod update;
 
 pub static SPRITE_SHEET_FOLDER: &str = "sprite-sheets";
 
@@ -32,14 +28,25 @@ pub mod prelude {
     pub use crate::{
         animations::AnimationSystems,
         get_decompressed_file_name,
-        material::{embed_sprite_assets, initialize_sprite_material, SpriteMaterial},
-        meshes::{RectMeshes, SpriteMeshes},
-        params::{SpriteOutline, SpriteParams},
+        loading::{
+            loaded::{LoadedAppearance, LoadedAppearances, LoadedSprite},
+            systems::{
+                load_from_entities_system, load_sprite_system, process_load_events_system,
+                store_loaded_appearances_system,
+            },
+            LoadAppearanceEvent,
+        },
+        material::{
+            embed_sprite_assets, initialize_sprite_material,
+            meshes::{RectMeshes, SpriteMeshes},
+            params::{SpriteOutline, SpriteParams},
+            SpriteMaterial,
+        },
         sprite_animations::*,
-        sprites::*,
+        update::update_sprite_system,
         SpriteSystems, SPRITE_SHEET_FOLDER,
     };
 
     #[cfg(feature = "debug")]
-    pub use crate::sprites::debug_sprites;
+    pub use crate::loading::debug::debug_sprites;
 }
