+++ crates/ryot_sprites/src/loading/systems.rs
@@ -1,79 +1,20 @@
-//! Sprite loading and drawing.
+use crate::material::meshes::SpriteMeshes;
+use crate::material::SpriteMaterial;
 use crate::prelude::*;
-use bevy_asset::*;
-use bevy_ecs::prelude::*;
-use bevy_render::prelude::*;
-use bevy_sprite::Mesh2dHandle;
+use bevy_asset::{AssetServer, Assets, Handle};
+use bevy_ecs::change_detection::{Res, ResMut};
+use bevy_ecs::event::{EventReader, EventWriter};
+use bevy_ecs::prelude::{Changed, In, Or, Query};
+use bevy_render::prelude::Image;
 use bevy_utils::tracing::warn;
-use bevy_utils::*;
-use derive_more::*;
+use bevy_utils::{default, HashMap, HashSet};
 use itertools::Itertools;
-use ryot_core::prelude::*;
-use ryot_tiled::prelude::*;
+use ryot_core::content::{ContentId, ContentType, VisualElements};
+use ryot_core::prelude::{FrameGroup, SpriteInfo, SpriteSheets, TextureAtlasLayouts};
+use ryot_tiled::tile_size;
 use std::path::PathBuf;
 
-#[cfg(feature = "debug")]
-use bevy_hierarchy::prelude::*;
-#[cfg(feature = "debug")]
-use bevy_sprite::Anchor;
-#[cfg(feature = "debug")]
-use bevy_stroked_text::{StrokedText, StrokedTextBundle};
-#[cfg(feature = "debug")]
-use bevy_transform::prelude::Transform;
-#[cfg(feature = "debug")]
-use glam::Vec3;
-
-pub struct LoadedAppearance {
-    pub sprites: Vec<LoadedSprite>,
-    pub layers: u32,
-    pub animation: Option<(AnimationKey, AnimationDescriptor)>,
-}
-
-#[derive(Resource, Default, Deref, DerefMut)]
-pub struct LoadedAppearances(pub HashMap<(ContentId, FrameGroup), LoadedAppearance>);
-
-/// A struct that holds the information needed to draw a sprite.
-/// It's a wrapper around a sprite sheet and a sprite id, that also holds the
-/// handle to the texture atlas.
-#[derive(Debug, Clone)]
-pub struct LoadedSprite {
-    pub sprite_id: u32,
-    pub group: ContentType,
-    pub sprite_sheet: SpriteSheet,
-    pub texture: Handle<Image>,
-    pub material: Handle<SpriteMaterial>,
-    pub mesh: Handle<Mesh>,
-}
-
-impl LoadedSprite {
-    pub fn new(
-        group: ContentType,
-        sprite_id: u32,
-        sprite_sheets: &SpriteSheets,
-        textures: &HashMap<String, Handle<Image>>,
-        material: &Handle<SpriteMaterial>,
-        mesh: &Handle<Mesh>,
-    ) -> Option<Self> {
-        let sprite_sheet = sprite_sheets.get_by_sprite_id(sprite_id)?;
-        let texture = textures.get(&sprite_sheet.file)?;
-        Some(Self {
-            group,
-            sprite_id,
-            sprite_sheet: sprite_sheet.clone(),
-            texture: texture.clone(),
-            material: material.clone(),
-            mesh: mesh.clone(),
-        })
-    }
-
-    pub fn get_sprite_index(&self) -> usize {
-        self.sprite_sheet
-            .get_sprite_index(self.sprite_id)
-            .expect("Sprite must exist in sheet")
-    }
-}
-
-pub fn load_sprites(
+fn load_sprites(
     group: ContentType,
     sprite_ids: Vec<u32>,
     sprite_sheets: &Res<SpriteSheets>,
@@ -124,150 +65,6 @@ pub fn load_sprites(
         .collect()
 }
 
-pub fn load_sprite_textures(
-    sprite_ids: Vec<u32>,
-    asset_server: &Res<AssetServer>,
-    sprite_sheets: &SpriteSheets,
-) -> Vec<(u32, Handle<Image>)> {
-    sprite_ids
-        .iter()
-        .filter_map(|sprite_id| {
-            load_sprite_texture(*sprite_id, asset_server, sprite_sheets)
-                .map(|texture| (*sprite_id, texture))
-        })
-        .collect()
-}
-
-pub fn load_sprite_texture(
-    sprite_id: u32,
-    asset_server: &Res<AssetServer>,
-    sprite_sheets: &SpriteSheets,
-) -> Option<Handle<Image>> {
-    let Some(sprite_sheet) = sprite_sheets.get_by_sprite_id(sprite_id) else {
-        warn!("Sprite {} not found in sprite sheets", sprite_id);
-        return None;
-    };
-
-    let texture: Handle<Image> = asset_server.load(
-        PathBuf::from(SPRITE_SHEET_FOLDER).join(get_decompressed_file_name(&sprite_sheet.file)),
-    );
-
-    Some(texture)
-}
-
-/// A system that ensures that all entities with an ContentId have a SpriteMaterial mesh bundle.
-#[cfg(feature = "debug")]
-pub fn debug_sprites(
-    mut commands: Commands,
-    q_debug: Query<(Entity, &Layer), (With<ContentId>, Without<Handle<SpriteMaterial>>)>,
-) {
-    q_debug.iter().for_each(|(entity, layer)| {
-        commands.entity(entity).with_children(|builder| {
-            builder.spawn((
-                StrokedTextBundle::new(StrokedText {
-                    font_size: 16.,
-                    text_anchor: Anchor::BottomRight,
-                    ..default()
-                })
-                .with_transform(
-                    Transform::from_translation(Vec3::new(8., debug_y_offset(layer), 1.))
-                        .with_scale(Vec3::splat(0.18)),
-                ),
-                PositionDebugText,
-                Layer::Hud(0),
-            ));
-            builder.spawn((
-                StrokedTextBundle::new(StrokedText {
-                    text: format!("{}", layer),
-                    font_size: 16.,
-                    text_anchor: Anchor::BottomLeft,
-                    color: Color::from(layer),
-                    ..default()
-                })
-                .with_transform(
-                    Transform::from_translation(Vec3::new(8.5, debug_y_offset(layer), 1.))
-                        .with_scale(Vec3::splat(0.18)),
-                ),
-                Layer::Hud(0),
-            ));
-        });
-    });
-}
-
-pub type ChangingAppearanceFilter = Or<(
-    Changed<ContentId>,
-    Changed<FrameGroup>,
-    Changed<Directional>,
-    (Without<AnimationSprite>, Changed<SpriteParams>),
-)>;
-
-pub fn update_sprite_system(
-    mut q_updated: Query<
-        (
-            &ContentId,
-            Option<&FrameGroup>,
-            Option<&Directional>,
-            Option<&SpriteParams>,
-            &mut SpriteLayout,
-            &mut Mesh2dHandle,
-            &mut Handle<SpriteMaterial>,
-        ),
-        ChangingAppearanceFilter,
-    >,
-    mut materials: ResMut<Assets<SpriteMaterial>>,
-    loaded_appereances: Res<LoadedAppearances>,
-) {
-    q_updated.iter_mut().for_each(
-        |(object_id, frame_group, direction, sprite_params, mut layout, mut mesh, mut material)| {
-            if object_id.is_none() {
-                return;
-            }
-            let Some(loaded_appearance) =
-                loaded_appereances.get(&(*object_id, frame_group.cloned().unwrap_or_default()))
-            else {
-                warn!("BUG: Loaded appearance for {:?} not found.", object_id);
-                return;
-            };
-
-            let direction_index = match direction {
-                Some(dir) => dir.index(),
-                None => 0,
-            } * loaded_appearance.layers as usize;
-
-            let Some(sprite) = loaded_appearance.sprites.get(direction_index) else {
-                warn!(
-                    "Sprite for appearance {:?} not found for direction {:?}",
-                    object_id, direction
-                );
-                return;
-            };
-            *layout = sprite.sprite_sheet.layout;
-            *mesh = Mesh2dHandle(sprite.mesh.clone());
-            *material = sprite_material_from_params(sprite_params, &mut materials, sprite);
-        },
-    );
-}
-
-pub fn sprite_material_from_params(
-    sprite_params: Option<&SpriteParams>,
-    materials: &mut ResMut<'_, Assets<SpriteMaterial>>,
-    sprite: &LoadedSprite,
-) -> Handle<SpriteMaterial> {
-    sprite_params
-        .map(|params| {
-            if params.has_any() {
-                materials
-                    .get(sprite.material.id())
-                    .map(|base| params.to_material(base.clone()))
-                    .map(|material| materials.add(material))
-                    .unwrap_or_else(|| sprite.material.clone())
-            } else {
-                sprite.material.clone()
-            }
-        })
-        .unwrap_or_else(|| sprite.material.clone())
-}
-
 pub fn load_from_entities_system(
     query: Query<(&ContentId, Option<&FrameGroup>), Or<(Changed<ContentId>, Changed<FrameGroup>)>>,
     loaded_appearances: Res<LoadedAppearances>,
@@ -285,12 +82,6 @@ pub fn load_from_entities_system(
                 frame_group: frame_group.cloned().unwrap_or_default(),
             });
         });
-}
-
-#[derive(Event)]
-pub struct LoadAppearanceEvent {
-    pub object_id: ContentId,
-    pub frame_group: FrameGroup,
 }
 
 pub fn process_load_events_system(
@@ -426,4 +217,35 @@ pub fn store_loaded_appearances_system(
 
             loaded_appearances.insert((*object_id, *frame_group), loaded_appearance);
         });
+}
+
+fn load_sprite_textures(
+    sprite_ids: Vec<u32>,
+    asset_server: &Res<AssetServer>,
+    sprite_sheets: &SpriteSheets,
+) -> Vec<(u32, Handle<Image>)> {
+    sprite_ids
+        .iter()
+        .filter_map(|sprite_id| {
+            load_sprite_texture(*sprite_id, asset_server, sprite_sheets)
+                .map(|texture| (*sprite_id, texture))
+        })
+        .collect()
+}
+
+fn load_sprite_texture(
+    sprite_id: u32,
+    asset_server: &Res<AssetServer>,
+    sprite_sheets: &SpriteSheets,
+) -> Option<Handle<Image>> {
+    let Some(sprite_sheet) = sprite_sheets.get_by_sprite_id(sprite_id) else {
+        warn!("Sprite {} not found in sprite sheets", sprite_id);
+        return None;
+    };
+
+    let texture: Handle<Image> = asset_server.load(
+        PathBuf::from(SPRITE_SHEET_FOLDER).join(get_decompressed_file_name(&sprite_sheet.file)),
+    );
+
+    Some(texture)
 }
