+++ crates/ryot_sprites/src/material/meshes.rs
